Deprecate old unicode utils
